BE/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
psycopg-binary
pandas
openpyxl
pytz
Pillow
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0